+++ color-manager/src/main/java/com/my_wall_color/color_manager/adapter/JpaUserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface JpaUserRepository extends JpaRepository<JpaUser, Long> {
+public interface JpaUserRepository extends JpaRepository<JpaUser, Integer> {
     Optional<JpaUser> findByUsername(String username);
 }
